requirements.txt
opencv-python
numpy
mss
pyautogui
psutil
pyyaml
pillow